# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A personal geek media tracker API - Track anime, movies, series, games, manga, books and music\nFeatures: Personal lists, progress tracking, ratings, favorites, and more",
        "title": "Geekery API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Rafael C Ribeiro",
            "email": "support@geekery.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/tags": {
            "get": {
                "description": "Get a list of all available tags",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Get all tags",
                "responses": {
                    "200": {
                        "description": "Success - returns all tags",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.Tag"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new tag for categorizing items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Create tag",
                "parameters": [
                    {
                        "description": "Tag data",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.CreateTagRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tag created successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.Tag"
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tags/{id}": {
            "get": {
                "description": "Get a specific tag by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Get tag by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - returns tag",
                        "schema": {
                            "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.Tag"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Tag not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing tag's data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Update tag",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated tag data",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.UpdateTagRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag updated successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.Tag"
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Tag not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a tag from the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Delete tag",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Tag not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_rafaelc-rb_geekery-api_internal_models.AnimeData": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "episodes": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "description": "FK para items",
                    "type": "integer"
                },
                "studio": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_rafaelc-rb_geekery-api_internal_models.BookData": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "chapters": {
                    "description": "Total de capítulos",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "description": "FK para items",
                    "type": "integer"
                },
                "pages": {
                    "description": "Total de páginas",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "volumes": {
                    "description": "Total de volumes (manga/LN sempre tem, books às vezes)",
                    "type": "integer"
                }
            }
        },
        "github_com_rafaelc-rb_geekery-api_internal_models.GameData": {
            "type": "object",
            "properties": {
                "average_playtime": {
                    "description": "Horas médias para completar",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "developer": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "description": "FK para items",
                    "type": "integer"
                },
                "platform": {
                    "description": "PC, PS5, Switch, Xbox, etc",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_rafaelc-rb_geekery-api_internal_models.Item": {
            "type": "object",
            "properties": {
                "anime_data": {
                    "description": "Dados específicos por tipo (apenas um será não-nil baseado no Type)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.AnimeData"
                        }
                    ]
                },
                "book_data": {
                    "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.BookData"
                },
                "cover_url": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "external_metadata": {
                    "description": "Metadados de APIs externas (MAL, IMDb, etc)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.JSONB"
                        }
                    ]
                },
                "game_data": {
                    "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.GameData"
                },
                "id": {
                    "type": "integer"
                },
                "movie_data": {
                    "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.MovieData"
                },
                "music_data": {
                    "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.MusicData"
                },
                "release_date": {
                    "description": "Data de lançamento/estreia",
                    "type": "string"
                },
                "series_data": {
                    "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.SeriesData"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.Tag"
                    }
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.MediaType"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_rafaelc-rb_geekery-api_internal_models.JSONB": {
            "type": "object",
            "additionalProperties": true
        },
...
        "github_com_rafaelc-rb_geekery-api_internal_models.MediaType": {
            "type": "string",
            "enum": [
                "anime",
                "movie",
                "series",
                "game",
                "manga",
                "light_novel",
                "music",
                "book"
            ],
            "x-enum-varnames": [
                "MediaTypeAnime",
                "MediaTypeMovie",
                "MediaTypeSeries",
                "MediaTypeGame",
                "MediaTypeManga",
                "MediaTypeLightNovel",
                "MediaTypeMusic",
                "MediaTypeBook"
            ]
        },
        "github_com_rafaelc-rb_geekery-api_internal_models.MovieData": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "director": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "description": "FK para items",
                    "type": "integer"
                },
                "runtime": {
                    "description": "em minutos",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_rafaelc-rb_geekery-api_internal_models.MusicData": {
            "type": "object",
            "properties": {
                "album": {
                    "type": "string"
                },
                "artist": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "duration": {
                    "description": "Duração total do álbum em segundos",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "description": "FK para items",
                    "type": "integer"
                },
                "tracks": {
                    "description": "Número de faixas",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "github_com_rafaelc-rb_geekery-api_internal_models.SeriesData": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "episodes": {
                    "description": "Total de episódios",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "description": "FK para items",
                    "type": "integer"
                },
                "seasons": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_rafaelc-rb_geekery-api_internal_models.Tag": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.Item"
                    }
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "internal_handlers.CreateTagRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.UpdateTagRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token (future implementation)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}